+++ macros/staging/get_collection_history_columns.sql
@@ -1,0 +1,25 @@
+{% macro get_collection_history_columns() %}
+
+{% set columns = [
+    {"name": "_fivetran_active", "datatype": "boolean"},
+    {"name": "_fivetran_start", "datatype": dbt.type_timestamp()},
+    {"name": "_fivetran_end", "datatype": dbt.type_timestamp()},
+    {"name": "_fivetran_synced", "datatype": dbt.type_timestamp()},
+    {"name": "created_at", "datatype": dbt.type_timestamp()},
+    {"name": "default_locale", "datatype": dbt.type_string()},
+    {"name": "description", "datatype": dbt.type_string()},
+    {"name": "help_center_id", "datatype": dbt.type_string()},
+    {"name": "icon", "datatype": dbt.type_string()},
+    {"name": "id", "datatype": dbt.type_string()},
+    {"name": "name", "datatype": dbt.type_string()},
+    {"name": "order", "datatype": dbt.type_int(), "quote": True, "alias": "display_order"},
+    {"name": "parent_id", "datatype": dbt.type_string()},
+    {"name": "type", "datatype": dbt.type_string()},
+    {"name": "updated_at", "datatype": dbt.type_timestamp()},
+    {"name": "url", "datatype": dbt.type_string()},
+    {"name": "workspace_id", "datatype": dbt.type_string()}
+] %}
+
+{{ return(columns) }}
+
+{% endmacro %}